#Work Examples
Code created for potential employers to review.
This repository includes examples of MySQL exercises I completed.  Instagram is from the last set of exercises in the Udemy course I took.  It includes the database used, the exercises assigned and my solutions as well as Colt Steele's solutions.  
employees_sample_database1 is from a web site for MySQL developers and designed to give developers more practice with their MySQL skills: https://dev.mysql.com/doc/employee/en/ .  There are no exercises provided with this site.  It is a database that allows developers to practice their coding skills.